// app/interfaces/register.ts
export interface Register {
  person_first_name: string;
  person_surname: string;
  institution_name: string;
  profile_id: string;
  institution_type: string;
  grading_system: string;
  institution_description: string;

  country: string;
  region: string;
  city: string;
  address: string;
  business_email: string;
  phone: string;
  website: string;
  instagram: string;
  facebook: string;
  twitter: string;
}

// Aquí puedes agregar otras interfaces relacionadas con el registro en el futuro